host_events: Bump up host events and masks to 64-bit
With the upcoming change to add a new command to get/set/clear host
events and masks, it seems to be the right time to bump up the host
events and masks to 64-bit. We are already out of available host
events. This change opens up at least 32 bits for new host events.

Old EC commands to operate on host events/masks will still deal with
lower 32-bits of the events/mask. On the other hand, the new command
being added will take care of the entire 64-bit events/masks. This
ensures that old BIOS and kernel versions can still work with the
newer EC versions.

BUG=b:69329196
BRANCH=None
TEST=make -j buildall. Verified:
1. hostevent set 0x4000 ==> Sets correct bit in host events
2. hostevent clear 0x4000 ==> Clears correct bit in host events
3. Kernel is able to query and read correct host event bits from
EC. Verified using evtest.
4. Coreboot is able to read correct wake reason from EC. Verified
using mosys eventlog list.

+++ include/host_command.h
@@ -150,14 +150,14 @@ uint16_t host_command_process(struct host_cmd_handler_args *args);
  *
  * @param mask          Event bits to set (use EC_HOST_EVENT_MASK()).
  */
-void host_set_events(uint32_t mask);
+void host_set_events(uint64_t mask);
 
 /**
  * Set a single host event.
  *
  * @param event         Event to set (EC_HOST_EVENT_*).
  */
-static inline void host_set_single_event(int event)
+static inline void host_set_single_event(uint64_t event)
 {
 	host_set_events(EC_HOST_EVENT_MASK(event));
 }
@@ -168,12 +168,12 @@ static inline void host_set_single_event(int event)
  * @param mask          Event bits to clear (use EC_HOST_EVENT_MASK()).
  *                      Write 1 to a bit to clear it.
  */
-void host_clear_events(uint32_t mask);
+void host_clear_events(uint64_t mask);
 
 /**
  * Return the raw event state.
  */
-uint32_t host_get_events(void);
+uint64_t host_get_events(void);
 
 /**
  * Check a single host event.
